requirements.txt
pillow==5.2.0